+++ askomics/libaskomics/OntologyManager.py
@@ -217,17 +217,20 @@ class OntologyManager(Params):
         list of dict
             Results
         """
+
+        max_results = self.settings.getint("askomics", "autocomplete_max_results", fallback=10)
+
         if ontology_type == "local":
             query = SparqlQuery(self.app, self.session, get_graphs=False)
             # TODO: Actually store the graph in the ontology to quicken search
             query.set_graphs([onto_graph])
-            return query.autocomplete_local_ontology(ontology_uri, query_term)
+            return query.autocomplete_local_ontology(ontology_uri, query_term, max_results)
         elif ontology_type == "ols":
             base_url = "https://www.ebi.ac.uk/ols/api/suggest"
             arguments = {
                 "q": query_term,
                 "ontology": quote_plus(onto_short_name.lower()),
-                "rows": 5
+                "rows": max_results
             }
 
             r = requests.get(base_url, params=arguments)