+++ askomics/react/src/routes/admin/ontologies.jsx
@@ -3,6 +3,7 @@ import axios from 'axios'
 import {Button, Form, FormGroup, Label, Input, Alert, Row, Col, CustomInput } from 'reactstrap'
 import BootstrapTable from 'react-bootstrap-table-next'
 import paginationFactory from 'react-bootstrap-table2-paginator'
+import update from 'react-addons-update'
 import PropTypes from 'prop-types'
 import Utils from '../../classes/utils'
 import { Redirect } from 'react-router-dom'
@@ -41,17 +42,35 @@ export default class Ontologies extends Component {
     return this.state.ontologiesSelected.length == 0
   }
 
+  cleanupOntologies (newOntologies) {
+    let cleanOntologies = []
+    newOntologies.map(onto => {
+      cleanOntologies.push({
+          id:onto.id,
+          name:onto.name,
+          uri:onto.uri,
+          short_name: onto.short_name,
+          type: onto.type
+      })
+    })
+    return cleanOntologies
+  }
+
   deleteSelectedOntologies () {
     let requestUrl = '/api/admin/delete_ontologies'
     let data = {
       ontologiesIdToDelete: this.state.ontologiesSelected
     }
+
     axios.post(requestUrl, data, { baseURL: this.props.config.proxyPath, cancelToken: new axios.CancelToken((c) => { this.cancelRequest = c }) })
       .then(response => {
         console.log(requestUrl, response.data)
         this.setState({
           ontologies: response.data.ontologies,
           ontologiesSelected: [],
+        })
+        this.props.setStateNavbar({
+          config: update(this.props.config, {ontologies: {$set: this.cleanupOntologies(response.data.ontologies)}})
         })
       })
   }
@@ -118,6 +137,9 @@ export default class Ontologies extends Component {
         uri: "",
         shortName: "",
         type: "local"
+      })
+      this.props.setStateNavbar({
+        config: update(this.props.config, {ontologies: {$set: this.cleanupOntologies(response.data.ontologies)}})
       })
     })
     .catch(error => {
@@ -322,6 +344,7 @@ export default class Ontologies extends Component {
 }
 
 Ontologies.propTypes = {
+  setStateNavbar: PropTypes.func,
   waitForStart: PropTypes.bool,
   config: PropTypes.object
 }